Update dependencies and footer copyright year calculation
This commit updates the development dependencies to their latest versions: parcel (2.10.3), prettier (3.1.1), rimraf (5.0.5), and sass (1.69.5). The project also now uses the "@parcel/config-default" and "@parcel/transformer-raw" dependencies with their respective versions (2.10.3). Additionally, the copyright year in the footer of the WordPasswords component is now dynamically calculated using JavaScript's `Date` object. This ensures that the year displayed in the footer always matches the current year. This change improves the project's compatibility, maintenance, and compliance with copyright conventions. No issues are associated with this commit.